+++ src/py_bugger/py_bugger.py
@@ -4,13 +4,28 @@ import random
 from pathlib import Path
 
 
+class ImportCollector(cst.CSTVisitor):
+    """Visit all import nodes, without modifying."""
+
+    def __init__(self):
+        self.import_nodes = []
+
+    def visit_Import(self, node):
+        """Collect all import nodes."""
+        self.import_nodes.append(node)
+
+
 class ImportModifier(cst.CSTTransformer):
     """Modify imports in the user's project."""
+
+    def __init__(self, nodes_to_break):
+        self.nodes_to_break = nodes_to_break
 
     def leave_Import(self, original_node, updated_node):
         """Modify a direct `import <package>` statement."""
         names = updated_node.names
-        if names:
+
+        if original_node in self.nodes_to_break:
             original_name = names[0].name.value
 
             # Remove one letter from the package name.
@@ -27,7 +42,11 @@ class ImportModifier(cst.CSTTransformer):
         return updated_node
 
 
-def main(exception_type, target_dir):
+def main(exception_type, target_dir, num_bugs):
+
+    # Set a random seed when testing.
+    if seed := os.environ.get("PY_BUGGER_RANDOM_SEED"):
+        random.seed(int(seed))
 
     if exception_type == "ModuleNotFoundError":
         print("Introducing a ModuleNotFoundError...")
@@ -46,8 +65,15 @@ def main(exception_type, target_dir):
         source = path.read_text()
         tree = cst.parse_module(source)
 
+        # Collect all import nodes.
+        import_collector = ImportCollector()
+        tree.visit(import_collector)
+        # breakpoint()
+
+        nodes_to_break = random.choices(import_collector.import_nodes, k=num_bugs)
+
         # Modify user's code.
-        modified_tree = tree.visit(ImportModifier())
+        modified_tree = tree.visit(ImportModifier(nodes_to_break))
 
         # Rewrite user's code.
         path.write_text(modified_tree.code)